// artifacts/api-server/src/sya/lib/schemas.ts
export const VerifyAccessCodeBody = zod.object({
  accessCode: zod.string(),
});